# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "KPI Laravel API",
        "description": "Multi-tenant HR KPI platform. Covers v1 (existing) and v2 (multi-tenant) endpoints.",
        "contact": {
            "email": "admin@system.local"
        },
        "version": "2.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Local / Development"
        }
    ],
    "paths": {
...
        "/api/v2/kpi/templates": {
            "get": {
                "tags": [
                    "KPI Templates"
                ],
                "summary": "List KPI templates",
                "operationId": "9a94a9cebcd88c04b239eaa5efe3c149",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "KPI Templates"
                ],
                "summary": "Create KPI template",
                "operationId": "079a27dfc90ea78e755f80c71700ffdb",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v2/kpi/templates/{id}": {
            "get": {
                "tags": [
                    "KPI Templates"
                ],
                "summary": "Get template with indicators",
                "operationId": "e225ec9fb1d05b8f404fae059b3cc7a2",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "KPI Templates"
                ],
                "summary": "Update KPI template",
                "operationId": "7cbd55a122308a3b03b31739eb21dd3b",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "KPI Templates"
                ],
                "summary": "Soft-delete KPI template",
                "operationId": "fa5e8ae9e0b1edb1e971d77b81ec07e4",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v2/kpi/templates/{id}/indicators": {
            "post": {
                "tags": [
                    "KPI Templates"
                ],
                "summary": "Add indicator to template",
                "operationId": "699670601d5cc73533fa54e2bebe9462",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v2/kpi/templates/{templateId}/indicators/{id}": {
            "put": {
                "tags": [
                    "KPI Templates"
                ],
                "summary": "Update indicator",
                "operationId": "f17eb88d67443bf8e1c81d6859a0c165",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "KPI Templates"
                ],
                "summary": "Remove indicator",
                "operationId": "adfda8583aea53e78f3ddc04dca2c634",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v2/tasks/v2": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "List tasks",
                "operationId": "bbb60afd56305d742176f1c0ad6b666d",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Create task",
                "operationId": "a87717531539b37beb1e4012ca87d9e3",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v2/tasks/v2/{id}": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Get task",
                "operationId": "1867c950f4feb6402273bdaed068d567",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Update task",
                "operationId": "ef1893e73f32eeb2448ccd6a07db73d0",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Delete task",
                "operationId": "bd034f49fd9752d64630c4bf58309f41",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v2/tasks/v2/mine": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Get my tasks",
                "operationId": "09122c0971e6291eec64e6a3f595d60d",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Include: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}